backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
DeepPurpose
pandas
numpy
git+https://github.com/bp-kelley/descriptastorus
sqlalchemy
rdkit-pypi
transformers